Quality of life improvements
Added an overload/specialisation that deduces non-callable template parameters. This allows a less verbose syntax for trivial for loops in a syntax that is more familiar to people

Minor fixes and improvements. Added a namespace, cleaned up some places, sprinkled some noexcept dust here and there, and tried to forward the loop_body (I have a bad feeling about this)

API is subject to changes in the future as it might not be possible to add some feautures that I want without distruptions or major changes